move api to core app; delete api app
