--- ohmg/api/tests.py
@@ -1,3 +1,0 @@
-from django.test import TestCase # noqa: F401
-
-# Create your tests here.
